backend/pyproject.toml
[project]
name = "core-discovery-api"
version = "0.2.0"
description = "CORE Discovery Framework - AI-powered product discovery coaching"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.18",
    "httpx>=0.28.0",
    "slowapi>=0.1.9",
    "pymupdf>=1.25.0",
    "python-pptx>=1.0.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
]

[project.optional-dependencies]
azure = [
    "openai>=1.60.0",
    "azure-identity>=1.19.0",
    "azure-storage-blob>=12.24.0",
    "azure-cosmos>=4.9.0",
    "azure-cognitiveservices-speech>=1.42.0",
    "PyJWT[crypto]>=2.9.0",
    "aiohttp>=3.10.0",
]
local = [
    "openai>=1.60.0",
    "aiosqlite>=0.20.0",
    "aiofiles>=24.1.0",
]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
